+++ src/main/java/edu/ie3/simona/api/data/ev/ExtEvDataConnection.java
@@ -15,21 +15,23 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.LinkedBlockingQueue;
-import org.apache.pekko.actor.ActorRef;
+import org.apache.pekko.actor.typed.ActorRef;
 
-public class ExtEvDataConnection implements ExtInputDataConnection {
+public class ExtEvDataConnection implements ExtInputDataConnection<EvDataMessageFromExt> {
   /** Data message queue containing messages from SIMONA */
   public final LinkedBlockingQueue<EvDataResponseMessageToExt> receiveTriggerQueue =
       new LinkedBlockingQueue<>();
 
   /** Actor reference to service that handles ev data within SIMONA */
-  private ActorRef dataService;
+  private ActorRef<EvDataMessageFromExt> dataService;
 
   /** Actor reference to adapter that handles scheduler control flow in SIMONA */
-  private ActorRef extSimAdapter;
+  private ActorRef<ScheduleDataServiceMessage<EvDataMessageFromExt>> extSimAdapter;
 
   @Override
-  public void setActorRefs(ActorRef dataService, ActorRef extSimAdapter) {
+  public void setActorRefs(
+      ActorRef<EvDataMessageFromExt> dataService,
+      ActorRef<ScheduleDataServiceMessage<EvDataMessageFromExt>> extSimAdapter) {
     this.dataService = dataService;
     this.extSimAdapter = extSimAdapter;
   }
@@ -99,9 +101,9 @@ public class ExtEvDataConnection implements ExtInputDataConnection {
    * @param msg the data/information that is sent to SIMONA's ev data service
    */
   public void sendExtMsg(EvDataMessageFromExt msg) {
-    dataService.tell(msg, ActorRef.noSender());
+    dataService.tell(msg);
     // we need to schedule data receiver activation with scheduler
-    extSimAdapter.tell(new ScheduleDataServiceMessage(dataService), ActorRef.noSender());
+    extSimAdapter.tell(new ScheduleDataServiceMessage<>(dataService));
   }
 
   /**