+++ src/main/java/edu/ie3/simona/api/data/primarydata/ExtPrimaryDataConnection.java
@@ -13,17 +13,17 @@ import edu.ie3.simona.api.data.primarydata.ontology.PrimaryDataMessageFromExt;
 import edu.ie3.simona.api.data.primarydata.ontology.ProvidePrimaryData;
 import java.util.*;
 import java.util.stream.Collectors;
-import org.apache.pekko.actor.ActorRef;
+import org.apache.pekko.actor.typed.ActorRef;
 import org.slf4j.Logger;
 
 /** Enables data connection of primary data between SIMONA and SimonaAPI */
-public class ExtPrimaryDataConnection implements ExtInputDataConnection {
+public class ExtPrimaryDataConnection implements ExtInputDataConnection<PrimaryDataMessageFromExt> {
 
   /** Actor reference to service that handles primary data within SIMONA */
-  private ActorRef dataService;
+  private ActorRef<PrimaryDataMessageFromExt> dataService;
 
   /** Actor reference to adapter that handles scheduler control flow in SIMONA */
-  private ActorRef extSimAdapter;
+  private ActorRef<ScheduleDataServiceMessage<PrimaryDataMessageFromExt>> extSimAdapter;
 
   /** Assets that provide primary data to SIMONA */
   private final Map<String, UUID> extPrimaryDataMapping;
@@ -33,7 +33,9 @@ public class ExtPrimaryDataConnection implements ExtInputDataConnection {
   }
 
   @Override
-  public void setActorRefs(ActorRef dataService, ActorRef extSimAdapter) {
+  public void setActorRefs(
+      ActorRef<PrimaryDataMessageFromExt> dataService,
+      ActorRef<ScheduleDataServiceMessage<PrimaryDataMessageFromExt>> extSimAdapter) {
     this.dataService = dataService;
     this.extSimAdapter = extSimAdapter;
   }
@@ -74,8 +76,8 @@ public class ExtPrimaryDataConnection implements ExtInputDataConnection {
    * @param msg the data/information that is sent to SIMONA's external primary data service
    */
   public void sendExtMsg(PrimaryDataMessageFromExt msg) {
-    dataService.tell(msg, ActorRef.noSender());
+    dataService.tell(msg);
     // we need to schedule data receiver activation with scheduler
-    extSimAdapter.tell(new ScheduleDataServiceMessage(dataService), ActorRef.noSender());
+    extSimAdapter.tell(new ScheduleDataServiceMessage<>(dataService));
   }
 }